Renamed nutrients usecase to trackedNutrients
