--- src/use_case/nutrients/NutrientsInputBoundary.java
@@ -1,5 +1,0 @@
-package src.use_case.nutrients;
-
-public interface NutrientsInputBoundary {
-    void execute(NutrientsInputData nutrientsInputData);
-}
