--- src/use_case/nutrients/NutrientsInteractor.java
@@ -1,6 +1,0 @@
-package src.use_case.nutrients;
-
-public class NutrientsInteractor implements NutrientsInputBoundary {
-    final NutrientsUserDataAccessInterface userDataAccessObject;  // name?
-    final NutrientsOutputBoundary nutrientsPresenter;
-}
