--- src/use_case/nutrients/NutrientsOutputBoundary.java
@@ -1,6 +1,0 @@
-package src.use_case.nutrients;
-
-public interface NutrientsOutputBoundary {
-    void prepareSuccessView(NutrientsOutputData nutrients);
-    void prepareFailView(String error);
-}
