+++ src/use_case/trackedNutrients/TrackedNutrientsInputBoundary.java
@@ -1,0 +1,5 @@
+package src.use_case.trackedNutrients;
+
+public interface TrackedNutrientsInputBoundary {
+    void execute(TrackedNutrientsInputData trackedNutrientsInputData);
+}
