+++ src/use_case/trackedNutrients/TrackedNutrientsOutputBoundary.java
@@ -1,0 +1,6 @@
+package src.use_case.trackedNutrients;
+
+public interface TrackedNutrientsOutputBoundary {
+    void prepareSuccessView(TrackedNutrientsOutputData nutrients);
+    void prepareFailView(String error);
+}
